Move theano.tensor.basic type definitions to theano.tensor.type
This change required a lot of general import refactoring in order to avoid (and
remove) circular dependency issues.

One major change involved the removal of imports like `from theano import
tensor`, because the module `tensor` clashes with the type `tensor` (e.g. `x =
tensor()`).  This usage was replaced with the conventional `import theano.tensor
as tt` or `from theano import tensor as tt` (the latter addresses a Python 3.6
import inconsistency/bug).  The same goes for `from theano import scalar`; the
name clashes with the `scalar` type name, so it was replaced with `import
theano.scalar as ts`.