Add Scan's inner-graph to its node inputs and use it for inner-graph compilation
This change adds `Scan`'s inner-graphs as "uncomputed" node inputs.  The inner
graphs inputs are converted to `NominalVariable`s and both the inner-inputs and
outputs are appended to the node (i.e. outer-)inputs.

When a `Scan`'s thunk is created, the inner-graph is extracted from the node
inputs and compiled with no optimizations (i.e. just lowered/transpiled and compiled).

In order for this to work, all uses of a `Scan`'s `node.inputs` needed to be
updated to `node.inputs\[:node.op.n_outer_inputs\]` so that the extra inner-graph
terms are ignored.