Change the update_storage parameter in CLazyLinker
This changes the `update_storage` parameter from a list containing
the input indices that are to be updated with the last N-many outputs
to a tuple of tuples specifying input/output indices.

Now, arbitrary output-to-input update pairings are possible, instead of
forcing graphs and code to compensate for this unnecessary restriction.