Downgrade flake8 pre-commit for py37 compatibility
Aesara's minimum-supported Python version is 3.7, and it's probably
really convenient to develop with that version in order to avoid using
incompatible language features. Then of course you want to run
pre-commit, but for that you either need a higher version of Python
(>=3.8.1) or a lower (<=5.0.4) version of flake8.

It seems to me like the most straightforward solution is to revert
flake8 to 5.0.4 (the second-most-recent release, August) until
Aesara drops support for Python 3.7.