Convert a test to use Hypothesis
This is a proof of concept for Hypothesis tests. I selected a test that
was easy to convert and replaced the existing test with hard-coded
values with a few tests that use Hypothesis.

Since it was pretty easy to do, I also added tests for complex inputs.
One thing I discovered is that cscalar doesn't work for any of the
numpy functions. Hypothesis inevitably finds a value that can't be
represented precisely and the test fails. All the functions work with
zscalar.

see: https://github.com/aesara-devs/aesara/issues/197